refactoring user input checks.
- rename independent prior to MultipleIndependent
- move user input checks and utils to separate module sbi/user_input/.
