+++ tests/linearGaussian_snl_test.py
@@ -9,7 +9,7 @@ from sbi.simulators.linear_gaussian import (
     get_true_posterior_samples_linear_gaussian_uniform_prior,
     linear_gaussian,
 )
-from sbi.simulators.user_input_checks import prepare_sbi_problem
+from sbi.user_input.user_input_checks import prepare_sbi_problem
 
 # use cpu by default
 torch.set_default_tensor_type("torch.FloatTensor")
